Bug 1222924 - Stop allowing webpages to link to moz-icon:
In UXP the implications for this security bug were changed so it actually didn't do what Mozilla intended in an attempt to preserve functionality internally. This did not completely work. This fix denies web access to any moz-icon but allows it to still work on other protocols like file: and about: etc.